# docs/openapi.yaml
openapi: 3.1.0
info:
  title: OpenDesk API
  description: >
    Sovereign, open-source office suite API. Provides document management,
    real-time collaboration, sharing, audit trails, workflow automation,
    cross-sovereign federation, local AI assistance, and verifiable data erasure.
  version: 0.1.0
  license:
    name: AGPL-3.0
    url: https://www.gnu.org/licenses/agpl-3.0.html
servers:
  - url: http://localhost:5500
    description: Local development server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: JWT or dev-mode token passed as Bearer header
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
      required: [error]
    ValidationError:
      type: object
      properties:
        error:
          type: string
        issues:
          type: array
          items:
            type: object
      required: [error, issues]
    OkResponse:
      type: object
      properties:
        ok:
          type: boolean
          const: true
      required: [ok]

    # --- Documents ---
...
    FederationPeer:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        endpointUrl:
          type: string
          format: uri
        publicKey:
          type: string
        trustLevel:
          type: string
          enum: [standard, elevated, restricted]
        status:
          type: string
          enum: [active, suspended, revoked]
        lastSeenAt:
          type: string
          nullable: true
        registeredBy:
          type: string
        createdAt:
          type: string
    RegisterPeerBody:
      type: object
      required: [name, endpointUrl, publicKey]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
        endpointUrl:
          type: string
          format: uri
        publicKey:
          type: string
          minLength: 1
        trustLevel:
          type: string
          enum: [standard, elevated, restricted]
          default: standard
    SendDocumentBody:
      type: object
      required: [documentId, peerId]
      properties:
        documentId:
          type: string
          format: uuid
        peerId:
          type: string
          format: uuid
    TransferBundle:
      type: object
      required: [sendingInstanceId, documentId, documentTitle, yjsStateBase64, signature, timestamp]
      properties:
        sendingInstanceId:
          type: string
        documentId:
          type: string
        documentTitle:
          type: string
        yjsStateBase64:
          type: string
        auditProofHash:
          type: string
        signature:
          type: string
        timestamp:
          type: string
    TransferRecord:
      type: object
      properties:
        id:
          type: string
          format: uuid
        peerId:
          type: string
          format: uuid
        direction:
          type: string
          enum: [inbound, outbound]
        documentId:
          type: string
        documentTitle:
          type: string
          nullable: true
        signature:
          type: string
        auditProofHash:
          type: string
          nullable: true
        status:
          type: string
          enum: [pending, completed, failed, rejected]
        error:
          type: string
          nullable: true
        createdAt:
          type: string

    # --- AI ---
...
security:
  - BearerAuth: []
paths:
...
  /api/federation/peers:
    get:
      tags: [Federation]
      summary: List federation peers
      description: List all registered federation peers.
      responses:
        "200":
          description: Array of peers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FederationPeer"
    post:
      tags: [Federation]
      summary: Register a federation peer
      description: Register a new peer instance for cross-sovereign federation. Admin only.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterPeerBody"
      responses:
        "201":
          description: Registered peer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FederationPeer"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationError"
  /api/federation/peers/{id}:
    patch:
      tags: [Federation]
      summary: Update peer status
      description: Update a peer's status (active, suspended, revoked). Admin only.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [active, suspended, revoked]
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
        "404":
          description: Peer not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/federation/send:
    post:
      tags: [Federation]
      summary: Send a document to a peer
      description: Initiate a signed document transfer to a federated peer. Admin only, requires manage permission on the document.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendDocumentBody"
      responses:
        "200":
          description: Transfer record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransferRecord"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationError"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/federation/receive:
    post:
      tags: [Federation]
      summary: Receive an inbound transfer
      description: Endpoint called by federated peers to deliver a signed document bundle. Authenticates via peer signature verification (no user auth).
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransferBundle"
      responses:
        "200":
          description: Transfer record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransferRecord"
        "400":
          description: Invalid transfer bundle
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationError"
        "401":
          description: Missing signature
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Unknown peer or invalid signature
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/federation/transfers:
    get:
      tags: [Federation]
      summary: List transfer history
      description: List federation transfer history, optionally filtered by peer.
      parameters:
        - name: peerId
          in: query
          schema:
            type: string
            format: uuid
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            maximum: 200
      responses:
        "200":
          description: Array of transfer records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TransferRecord"

  # ===== AI =====
...